updated setup, install more programs, make more generic

+++ vagrant_setup.sh
@@ -23,10 +23,13 @@ deb http://www.emdebian.org/debian squeeze main
 EOF
 sudo apt-get -y install libc6-mipsel-cross
 sudo apt-get -y install libc6-arm-cross
-mkdir /etc/qemu-binfmt
-ln -s /usr/mipsel-linux-gnu /etc/qemu-binfmt/mipsel 
-ln -s /usr/arm-linux-gnueabihf /etc/qemu-binfmt/arm
-rm /etc/apt/sources.list.d/emdebian.list
+
+if [ ! -d "/etc/qemu-binfmt" ]; then
+  mkdir /etc/qemu-binfmt
+  ln -s /usr/mipsel-linux-gnu /etc/qemu-binfmt/mipsel 
+  ln -s /usr/arm-linux-gnueabihf /etc/qemu-binfmt/arm
+  rm /etc/apt/sources.list.d/emdebian.list
+fi
 # sudo apt-get update
 
 # Install Binjitsu
@@ -42,36 +45,44 @@ git clone https://github.com/zachriggle/pwndbg
 echo source `pwd`/pwndbg/gdbinit.py >> ~/.gdbinit
 
 # Capstone for pwndbg
-git clone https://github.com/aquynh/capstone
-cd capstone
-git checkout -t origin/next
-sudo ./make.sh install
-cd bindings/python
-sudo python3 setup.py install # Ubuntu 14.04+, GDB uses Python3
+if [ ! -d "~/tools/capstone" ]; then
+  git clone https://github.com/aquynh/capstone
+  cd capstone
+  git checkout -t origin/next
+  sudo ./make.sh install
+  cd bindings/python
+  sudo python3 setup.py install # Ubuntu 14.04+, GDB uses Python3
+fi
 
 # pycparser for pwndbg
 sudo pip3 install pycparser # Use pip3 for Python3
 
 # Install radare2
-git clone https://github.com/radare/radare2
-cd radare2
-./sys/install.sh
+if [ ! -d "/usr/share/radare2" ]; then
+  git clone https://github.com/radare/radare2
+  cd radare2
+  ./sys/install.sh
+fi
 
 # Install binwalk
 cd 
-git clone https://github.com/devttys0/binwalk
-cd binwalk
-sudo python setup.py install
+if [ ! -d "~/binwalk" ]; then
+  git clone https://github.com/devttys0/binwalk
+  cd binwalk
+  sudo python setup.py install
+fi
 
 # Install Firmware-Mod-Kit
 sudo apt-get -y install git build-essential zlib1g-dev liblzma-dev python-magic
-cd ~/tools
-wget https://firmware-mod-kit.googlecode.com/files/fmk_099.tar.gz
-tar xvf fmk_099.tar.gz
-rm fmk_099.tar.gz
-cd fmk_099/src
-./configure
-make
+if [ ! -d "~/tools/fmk" ]; then
+  cd ~/tools
+  wget https://firmware-mod-kit.googlecode.com/files/fmk_099.tar.gz
+  tar xvf fmk_099.tar.gz
+  rm fmk_099.tar.gz
+  cd fmk_099/src
+  ./configure
+  make
+fi
 
 # Uninstall capstone
 sudo pip2 uninstall capstone -y
@@ -95,3 +106,35 @@ sudo pip install virtualenv
 virtualenv angr
 source angr/bin/activate
 pip install angr --upgrade
+
+# Install ropgadget
+if [ ! -d "~/tools/ROPgadget" ]; then
+  cd ~/tools
+  git clone https://github.com/JonathanSalwan/ROPgadget
+  cd ROPgadget
+  sudo python setup.py install
+fi
+
+# Install zsh
+sudo apt-get -y install zsh
+
+# Install oh-my-zsh
+sh -c "$(curl -fsSL https://raw.github.com/robbyrussell/oh-my-zsh/master/tools/install.sh)"
+
+# Install z.sh
+cd ~
+wget https://raw.githubusercontent.com/rupa/z/master/z.sh -O .z.sh
+
+# Update .zshrc
+# TODO this will be handled by stow
+cd
+wget https://raw.githubusercontent.com/CarlNeuhaus/config_files/master/zsh/zshrc -O .zshrc
+
+# Install vundle into vim
+git clone https://github.com/VundleVim/Vundle.vim.git ~/.vim/bundle/Vundle.vim
+vim +PluginInstall +qall
+
+# Update .vimrc
+# TODO this will be handled by stow
+cd ~
+wget https://raw.githubusercontent.com/CarlNeuhaus/config_files/master/vim/vimrc -O .vimrc
